dialog for opprettelse av revurdering er lagt til i egen komponent
